% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avltree.R
\name{assoc}
\alias{assoc}
\title{Assoc(iate) a key-value pair with an existing AVL tree map.}
\usage{
assoc(tree, key, value)
}
\arguments{
\item{value}{}
}
\description{
Assoc(iate) a key-value pair with an existing AVL tree map.
}
