% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avltree.R
\name{contains}
\alias{contains}
\title{Determine the presence of a key in the AVL tree map.}
\usage{
contains(tree, key, test = `==`)
}
\arguments{
\item{test}{}
}
\description{
Determine the presence of a key in the AVL tree map.
}
