% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avltree.R
\docType{data}
\name{empty}
\alias{empty}
\title{Sentinel object representing an empty AVL tree map.}
\format{An object of class \code{AVLEmptyNode} of length 0.}
\usage{
empty
}
\description{
Sentinel object representing an empty AVL tree map.
}
\keyword{datasets}
