% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avltree.R
\name{lookup}
\alias{lookup}
\title{Look up the value associated with a key in an AVL tree map.}
\usage{
lookup(tree, key, default = NULL)
}
\arguments{
\item{test}{}
}
\description{
Look up the value associated with a key in an AVL tree map.
}
